feat: adjust TypeScript version range and disable cache
This commit adjusts the TypeScript version range in multiple
package.json files to be ">=4.8.0 <5.5.0". This change is made to
ensure compatibility with a wider range of TypeScript versions.

In addition, the turbo.json file is updated to disable caching for the
"prepare" script. This is done to ensure that the most recent changes
are always included when the script is run.